# agent-runtime-contracts/src/main/resources/openapi/agent-runtime-v2.yaml
openapi: 3.1.0
info:
  title: Agent Runtime API
  version: 2.0.0
  description: |
    Asynchronous Agent Runtime contract. All jobs return a bounded JSON object.
    Large input and output use immutable objects transferred outside job JSON.
    Live events can include provider-supplied reasoning summaries, visible agent text and tool
    activity. Raw hidden chain-of-thought is not part of this contract.
    The v2 contract is implemented alongside the backwards-compatible v1 API.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /v2
security:
  - bearerAuth: []
paths:
  /uploads:
    post:
      operationId: createInputUpload
      summary: Reserve a resumable upload for job input
      description: |
        Use this before creating a job when input contains a file or text larger than the inline
        instruction limit. Binary files and text files such as text/plain, text/markdown and
        application/json use the same upload flow. The returned objectId can only be referenced
        after upload completion.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadRequest'
      responses:
        '201':
          description: Upload reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadView'
        '400': {$ref: '#/components/responses/ClientError'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
  /uploads/{uploadId}:
    parameters:
      - $ref: '#/components/parameters/UploadId'
    head:
      operationId: getInputUploadOffset
      summary: Read current upload offset and state
      responses:
        '204':
          description: Upload state returned in headers
          headers:
            Upload-Offset:
              required: true
              schema: {type: integer, format: int64, minimum: 0}
            Upload-Length:
              required: true
              schema: {type: integer, format: int64, minimum: 1}
            Upload-State:
              required: true
              schema: {type: string, enum: [UPLOADING, READY, EXPIRED]}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: appendInputUploadChunk
      summary: Append one binary chunk at the exact current offset
      parameters:
        - $ref: '#/components/parameters/UploadOffset'
      requestBody:
        required: true
        content:
          application/offset+octet-stream:
            schema: {type: string, format: binary}
      responses:
        '204':
          description: Chunk stored
          headers:
            Upload-Offset:
              required: true
              schema: {type: integer, format: int64, minimum: 0}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
    delete:
      operationId: deleteInputUpload
      summary: Delete an unbound upload and its object
      responses:
        '204': {description: Upload deleted or already absent}
        '409': {$ref: '#/components/responses/Conflict'}
  /uploads/{uploadId}/complete:
    post:
      operationId: completeInputUpload
      summary: Verify size and SHA-256 and make the input object immutable
      parameters:
        - $ref: '#/components/parameters/UploadId'
      responses:
        '200':
          description: Immutable input object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectView'
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /jobs:
    post:
      operationId: createJob
      summary: Create or return one idempotent AI session/job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
            examples:
              structuredText:
                value:
                  idempotencyKey: summarize-article-123
                  jobKind: APPLICATION_WORK
                  taskType: STRUCTURED_GENERATION
                  execution:
                    vendorId: openai
                    model: gpt-example
                    mode: SUBSCRIPTION
                  input:
                    instruction: Summarize the article in Dutch.
                    objects: []
                  output:
                    resultSchema:
                      type: object
                      required: [text]
                      properties:
                        text: {type: string}
                      additionalProperties: false
                    artifacts: []
                  executionTimeoutSeconds: 3600
              transcription:
                value:
                  idempotencyKey: transcribe-audio-123
                  jobKind: APPLICATION_WORK
                  taskType: TRANSCRIPTION
                  execution:
                    vendorId: openai
                    model: transcription-model
                    mode: API
                  input:
                    instruction: Transcribe in Dutch and include timestamps.
                    objects:
                      - objectId: 00000000-0000-0000-0000-000000000001
                        name: audio
                        role: SOURCE
                  output:
                    artifacts:
                      - name: transcript
                        required: true
                        mimeTypes: [text/plain, application/json]
                        maxBytes: 1073741824
                  executionTimeoutSeconds: 3600
      responses:
        '202':
          description: Job accepted, or existing idempotent job returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobView'
        '400': {$ref: '#/components/responses/ClientError'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
    get:
      operationId: listJobs
      summary: Cursor-page jobs visible to the authenticated tenant
      parameters:
        - {in: query, name: status, schema: {$ref: '#/components/schemas/JobStatus'}}
        - {in: query, name: cursor, schema: {type: string, maxLength: 500}}
        - {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 100, default: 30}}
      responses:
        '200':
          description: Job page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobPage'
  /jobs/{jobId}:
    get:
      operationId: getJob
      summary: Get current job status
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Current job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobView'
        '404': {$ref: '#/components/responses/NotFound'}
  /jobs/{jobId}/result:
    get:
      operationId: getJobResult
      summary: Get bounded JSON result and output object metadata
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Immutable successful result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResultView'
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /jobs/{jobId}/attempts:
    get:
      operationId: listJobAttempts
      summary: List technical attempts, retry reasons, usage and cost
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Attempts in chronological order
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/AttemptView'}
        '404': {$ref: '#/components/responses/NotFound'}
  /jobs/{jobId}/events:
    get:
      operationId: getJobEvents
      summary: Cursor-page durable status and progress events
      parameters:
        - $ref: '#/components/parameters/JobId'
        - {in: query, name: afterSequence, schema: {type: integer, format: int64, minimum: 0}}
        - {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 500, default: 100}}
      responses:
        '200':
          description: Event page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEventPage'
        '404': {$ref: '#/components/responses/NotFound'}
  /jobs/{jobId}/event-stream:
    get:
      operationId: streamJobEvents
      summary: Follow job status and progress with Server-Sent Events
      description: |
        Optional alternative to polling. Last-Event-ID resumes after a disconnect. Visible agent
        text, tool activity and provider-supplied reasoning summary chunks may be sent as bounded
        LOG_MESSAGE events. Raw output-token streams, large result bytes and unavailable hidden
        chain-of-thought are not sent on this stream.
      parameters:
        - $ref: '#/components/parameters/JobId'
        - in: header
          name: Last-Event-ID
          required: false
          schema: {type: integer, format: int64, minimum: 0}
      responses:
        '200':
          description: SSE stream of JobEventView payloads
          content:
            text/event-stream:
              schema: {type: string}
        '404': {$ref: '#/components/responses/NotFound'}
  /jobs/{jobId}/objects/{objectId}/content:
    get:
      operationId: downloadJobObject
      summary: Stream one input or output object owned by the job tenant
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/ObjectId'
        - in: header
          name: Range
          required: false
          schema: {type: string, pattern: '^bytes=[0-9]*-[0-9]*$'}
      responses:
        '200':
          description: Complete object
          headers:
            ETag: {schema: {type: string}}
            Accept-Ranges: {schema: {type: string, const: bytes}}
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
        '206':
          description: Requested byte range
          headers:
            Content-Range: {schema: {type: string}}
            ETag: {schema: {type: string}}
            Accept-Ranges: {schema: {type: string, const: bytes}}
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
        '404': {$ref: '#/components/responses/NotFound'}
        '416':
          description: Requested range is not satisfiable
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /jobs/{jobId}/cancel:
    post:
      operationId: cancelJob
      summary: Cancel queued work or request cancellation of the active attempt
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Current job after cancellation request
          content:
            application/json:
              schema: {$ref: '#/components/schemas/JobView'}
        '404': {$ref: '#/components/responses/NotFound'}
  /jobs/{jobId}/content:
    delete:
      operationId: deleteJobContent
      summary: Schedule early deletion of content belonging to a terminal job
      description: |
        Deletes input objects, output artifacts and retained monitor logs. Job status, attempts,
        usage and costs remain available. Running or queued jobs cannot be purged.
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '202':
          description: Content deletion scheduled
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ContentDeletionView'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /usage/summary:
    get:
      operationId: getTenantUsageSummary
      summary: Aggregate usage and cost for the authenticated tenant
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/Until'
        - $ref: '#/components/parameters/GroupBy'
        - {in: query, name: taskType, schema: {$ref: '#/components/schemas/TaskType'}}
        - {in: query, name: vendorId, schema: {type: string, maxLength: 100}}
        - {in: query, name: model, schema: {type: string, maxLength: 160}}
        - {in: query, name: mode, schema: {$ref: '#/components/schemas/ExecutionMode'}}
      responses:
        '200':
          description: Aggregated rows
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UsageSummaryResponse'}
        '400': {$ref: '#/components/responses/ClientError'}
  /management/usage/summary:
    get:
      operationId: getManagementUsageSummary
      summary: Aggregate usage and cost across all tenants
      description: Requires the administrator role. Group by TENANT to produce a project overview.
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/Until'
        - $ref: '#/components/parameters/GroupBy'
        - {in: query, name: tenantId, schema: {type: string, maxLength: 100}}
        - {in: query, name: taskType, schema: {$ref: '#/components/schemas/TaskType'}}
        - {in: query, name: vendorId, schema: {type: string, maxLength: 100}}
        - {in: query, name: model, schema: {type: string, maxLength: 160}}
        - {in: query, name: mode, schema: {$ref: '#/components/schemas/ExecutionMode'}}
      responses:
        '200':
          description: Aggregated rows
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UsageSummaryResponse'}
        '400': {$ref: '#/components/responses/ClientError'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /management/prices:
    get:
      operationId: listPriceRates
      summary: List current and historical immutable price rates
      parameters:
        - {in: query, name: vendorId, schema: {type: string, maxLength: 100}}
        - {in: query, name: model, schema: {type: string, maxLength: 160}}
      responses:
        '200':
          description: Price rates
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PriceRateView'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      operationId: createPriceRate
      summary: Add a new effective-dated immutable price rate
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreatePriceRateRequest'}
      responses:
        '201':
          description: Price rate created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PriceRateView'}
        '400': {$ref: '#/components/responses/ClientError'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /workers/register:
    post:
      operationId: registerWorker
      summary: Register or refresh worker capabilities
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WorkerRegistrationRequest'}
      responses:
        '200':
          description: Registration accepted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WorkerView'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    UploadId:
      in: path
      name: uploadId
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    ClientError:
      description: Invalid request
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
    CreateUploadRequest:
      type: object
      additionalProperties: false
      required: [filename, mimeType, sizeBytes, sha256]
      properties:
        filename: {$ref: '#/components/schemas/Filename'}
        mimeType: {type: string, minLength: 1, maxLength: 160}
        sizeBytes: {type: integer, format: int64, minimum: 1, maximum: 2147483648}
        sha256: {$ref: '#/components/schemas/Sha256'}
    UploadView:
      type: object
      additionalProperties: false
      required: [uploadId, objectId, state, protocol, chunkSizeBytes, uploadUrl, offset, sizeBytes, expiresAt]
      properties:
        uploadId: {type: string, format: uuid}
        objectId: {type: string, format: uuid}
        state: {type: string, enum: [UPLOADING, READY, EXPIRED]}
        protocol: {type: string, const: RESUMABLE_PATCH}
        chunkSizeBytes: {type: integer, format: int64, minimum: 1048576, maximum: 67108864}
        uploadUrl: {type: string}
        offset: {type: integer, format: int64, minimum: 0}
        sizeBytes: {type: integer, format: int64, minimum: 1}
        expiresAt: {type: string, format: date-time}
...
    Sha256:
      type: string
      pattern: '^[0-9a-f]{64}$'
...